backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyMuPDF==1.24.5
nltk==3.8.1
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
mlxtend==0.23.1
requests==2.32.3
beautifulsoup4==4.12.3
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.0
joblib==1.4.2
aiofiles==23.2.1
lxml==5.2.2
